Apply valid improvements from third code review
bf16 on A100 (ZeRO.py, Megatron.py):
  fp16 → bf16 across both benchmarks. On A100, bf16 is native hardware
  support. DeepSpeed's BF16Optimizer skips the fp32 master-weight copy
  that fp16 mode maintains, saving ~4 bytes/param (~6.65 GB/GPU for 6.7B).
  No loss scaling needed. Megatron TransformerConfig and model.bfloat16()
  updated to match.

ZeRO.py — wall_clock_breakdown and cpu_checkpointing:
  wall_clock_breakdown: True  gives per-step forward/backward/comm timing
  in DeepSpeed logs at no runtime cost — useful for analysis.
  cpu_checkpointing: False  is already the default but now explicit;
  prevents accidental CPU offload of checkpointed activations.

run_zero_bench.slurm — ntasks-per-node 4→1, cpus-per-task 4→16:
  torchrun manages its own worker processes internally; the correct Slurm
  pattern is ntasks=1 (one launcher task) with enough CPUs for all workers.
  ntasks=4 has Slurm assign one GPU per task slot which fights with
  torchrun's LOCAL_RANK-based device assignment. Total CPU count unchanged
  (1×16 = 4×4 = 16 CPUs).

NCCL_DEBUG=WARN added to all three slurm scripts:
  INFO floods the log with thousands of handshake lines per step.
  WARN captures actual errors and timeouts without the noise.

Skipped from feedback:
  - NCCL_SOCKET_IFNAME / NET_GDR_LEVEL: Delta-specific interface names
    unknown; wrong value silently breaks all NCCL communication.
  - overlap_comm: intentionally False for clean isolated measurement.
  - device_id removal: valid PyTorch 2.0+ API, confirmed by dry runs.
  - TRITON_CACHE_DIR: Delta GPU nodes have 1.5-2 TB node-local NVMe;
    /tmp is the right place for compiled kernel cache.

+++ Megatron.py
@@ -65,9 +65,10 @@ def _build_model(model_cfg: dict, seq_len: int) -> GPTModel:
         num_attention_heads=model_cfg["n_head"],
         ffn_hidden_size=4 * model_cfg["n_embd"],
         use_cpu_initialization=True,   # shard params at init, avoid GPU OOM
-        fp16=True,
-        params_dtype=torch.float16,
-        pipeline_dtype=torch.float16,
+        # bf16 is native on A100: no loss scaling, no fp32 master-weight copy.
+        bf16=True,
+        params_dtype=torch.bfloat16,
+        pipeline_dtype=torch.bfloat16,
         add_bias_linear=True,
         # Disable fused kernels — they require TransformerEngine / APEX
         bias_activation_fusion=False,
@@ -92,7 +93,7 @@ def _build_model(model_cfg: dict, seq_len: int) -> GPTModel:
         post_process=mpu.is_pipeline_last_stage(),
     )
 
-    return model.cuda().half()
+    return model.cuda().bfloat16()
 
 
 # ── Data iterator ─────────────────────────────────────────────────────────────
